ptbcontrib/redis_persistence/requirements.txt
redis-py
python-telegram-bot>=12.0